channel config is now working
I got the channel config to work properly. This is nice because there is now the ability to set the number of audio channels for an endpoint. IDK how this is exactly supposed to work just yt as I haven't messed around with it yet. But I can correctly collect the channels.