Adds support for changing a sessions default endpoint Adds support for enabling and disabling the audio enhancements ("Enhance Audio") for an endpoint
I have also reworked the IProperty store interface and the PROPVARIANT structure. I have to still do some more work on it.

I wrote a program that went through all of the files in the system32 folder and extracted any exported functions contained within them. I added some of the functions that might be useful for this library. I have not worked them into the library yet so that needs to be done. I also went through some of the Windows SDK idl files and found some additional interfaces that may be of use. I have ported the interfaces to python but I have not added them for use as of yet.